refactor:Attraction Repository 수정 pk값 변경

+++ src/main/java/com/example/planergram/repository/AttractionRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AttractionRepository extends JpaRepository<Attraction, Integer> {
+public interface AttractionRepository extends JpaRepository<Attraction, Long> {
     List<Attraction> findByAttractionAddress();
 }
